# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Regieren AI assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  deviceModel?: string;
  deviceProcessor?: string;
  rootStatus?: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface User {
  userId: string;
  email: string;
  plan: string;
  deviceModel?: string | null;
  deviceProcessor?: string | null;
  rootStatus?: string | null;
  telegramUsername?: string | null;
  telegramLinked?: boolean;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface ChatInput {
  /** @minLength 1 */
  message: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface Claim {
  serial: string;
  fileName: string;
  fileId?: string | null;
  claimedAt: string;
}

export interface RedeemInput {
  /** @minLength 1 */
  serial: string;
}

export interface TelegramLink {
  token: string;
  botConfigured: boolean;
}

export interface TelegramWebhookInput { [key: string]: unknown }

export interface Message {
  message: string;
}

export interface Error {
  error: string;
}

/**
 * Error
 */
export type ErrorResponse = Error;