# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Regieren AI assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
  - name: chat
  - name: claims
  - name: telegram
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /v1/auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a Regieren AI account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to a Regieren AI account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/Error"
  /v1/auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security: [{ bearerAuth: [] }]
      summary: Get the current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Error"
  /v1/auth/logout:
    post:
      operationId: logout
      tags: [auth]
      security: [{ bearerAuth: [] }]
      summary: Sign out the current user
      responses:
        "200":
          description: Signed out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /v1/chat:
    post:
      operationId: sendChat
      tags: [chat]
      security: [{ bearerAuth: [] }]
      summary: Send a message and receive an assistant response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "401":
          $ref: "#/components/responses/Error"
  /v1/chat/stream:
    post:
      operationId: streamChat
      tags: [chat]
      security: [{ bearerAuth: [] }]
      summary: Stream an assistant response as server-sent events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Text/event-stream response
          content:
            text/event-stream:
              schema:
                type: string
  /v1/chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      security: [{ bearerAuth: [] }]
      summary: Get recent assistant messages
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
  /v1/claims:
    get:
      operationId: listClaims
      tags: [claims]
      security: [{ bearerAuth: [] }]
      summary: List claims for the current user
      responses:
        "200":
          description: Claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Claim"
  /v1/claims/redeem:
    post:
      operationId: redeemClaim
      tags: [claims]
      security: [{ bearerAuth: [] }]
      summary: Redeem a premium file serial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemInput"
      responses:
        "200":
          description: Redeemed file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "400":
          $ref: "#/components/responses/Error"
  /v1/telegram/link:
    get:
      operationId: getTelegramLink
      tags: [telegram]
      security: [{ bearerAuth: [] }]
      summary: Get a Telegram linking token
      responses:
        "200":
          description: Link token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramLink"
  /v1/telegram/unlink:
    post:
      operationId: unlinkTelegram
      tags: [telegram]
      security: [{ bearerAuth: [] }]
      summary: Remove the current Telegram link
      responses:
        "200":
          description: Telegram unlinked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /v1/telegram/webhook:
    post:
      operationId: telegramWebhook
      tags: [telegram]
      summary: Receive a Telegram bot update
      parameters:
        - name: X-Telegram-Bot-Api-Secret-Token
          in: header
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramWebhookInput"
      responses:
        "200":
          description: Webhook accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        deviceModel: { type: string }
        deviceProcessor: { type: string }
        rootStatus: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
    User:
      type: object
      required: [userId, email, plan]
      properties:
        userId: { type: string }
        email: { type: string, format: email }
        plan: { type: string }
        deviceModel: { type: string, nullable: true }
        deviceProcessor: { type: string, nullable: true }
        rootStatus: { type: string, nullable: true }
        telegramUsername: { type: string, nullable: true }
        telegramLinked: { type: boolean }
    ChatInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1 }
    ChatMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id: { type: string }
        role: { type: string, enum: [user, assistant] }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    Claim:
      type: object
      required: [serial, fileName, claimedAt]
      properties:
        serial: { type: string }
        fileName: { type: string }
        fileId: { type: string, nullable: true }
        claimedAt: { type: string, format: date-time }
    RedeemInput:
      type: object
      required: [serial]
      properties:
        serial: { type: string, minLength: 1 }
    TelegramLink:
      type: object
      required: [token, botConfigured]
      properties:
        token: { type: string }
        botConfigured: { type: boolean }
    TelegramWebhookInput:
      type: object
      additionalProperties: true
    Message:
      type: object
      required: [message]
      properties:
        message: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer