# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Regieren AI assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatInput {
  /** @minLength 1 */
  message: string;
}