# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Regieren AI assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessageRole } from './chatMessageRole';

export interface ChatMessage {
  id: string;
  role: ChatMessageRole;
  content: string;
  createdAt: Date;
}