# lib/api-zod/src/generated/types/claim.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Regieren AI assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Claim {
  serial: string;
  fileName: string;
  fileId?: string | null;
  claimedAt: Date;
}