# lib/api-zod/src/generated/types/redeemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Regieren AI assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface RedeemInput {
  /** @minLength 1 */
  serial: string;
}