# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Regieren AI assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  email: string;
  /** @minLength 8 */
  password: string;
  deviceModel?: string;
  deviceProcessor?: string;
  rootStatus?: string;
}