# lib/api-zod/src/generated/types/telegramLink.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Regieren AI assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface TelegramLink {
  token: string;
  botConfigured: boolean;
}