# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Regieren AI assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  userId: string;
  email: string;
  plan: string;
  deviceModel?: string | null;
  deviceProcessor?: string | null;
  rootStatus?: string | null;
  telegramUsername?: string | null;
  telegramLinked?: boolean;
}